+++ lib/src/facade/amap_view.widget.dart
@@ -17,7 +17,7 @@ import 'enums.dart';
 import 'models.dart';
 
 part 'amap_controller.dart';
-part 'delegates.dart';
+part 'map_delegates.dart';
 
 typedef Future<void> _OnMapCreated(AmapController controller);
